New style for work cards, some formatting adjustments, file tree additions
Work cards have a new look, but are not finished yet, going to add a button that overlays a corresponding floating page for each project.

Format changing for headers, smaller margin bottom and top.

Added new folders for projectsLarge (holds components for each project page that is going to be displayed later)